Check Database Connection

This function checks if the provided connection is a valid database connection. The function will throw an error if the connection is not a valid DBI or pool connection.

Usage

check_db_conn(conn, arg = rlang::caller_arg(conn), call = rlang::caller_env())

check_db_pool(conn, arg = rlang::caller_arg(conn), call = rlang::caller_env())

Arguments

conn

A database connection object.

arg

An argument name as a string. This argument will be mentioned in error messages as the input that is at the origin of a problem.

call

The execution environment of a currently running function, e.g. caller_env(). The function will be mentioned in error messages as the source of the error. See the call argument of abort() for more information.

Value

The provided connection object invisibly if it is valid.