Create a Leasing Calendar Schedule Table

This function generates a leasing calendar schedule table for a given start year and number of years. The leasing season is defined as starting on September 1st and ending on August 1st of the following year.

The table includes the following columns:

date: The date in the sequence.

calendar_year: The calendar year of the date.

leasing_year: The leasing year corresponding to the date.

leasing_season_end_date: The end date of the leasing season for the leasing year.

leasing_season_start_date: The start date of the leasing season for the leasing year.

leasing_week: The week number within the leasing season.

weeks_left_to_lease: The number of weeks left to lease until the end of the leasing season.

is_current_leasing_season: A logical indicating if the date is within the current leasing season.

Usage

create_leasing_calendar(start_year = 2023, num_years = 3)

Arguments

start_year

The starting year for the leasing calendar. Defaults to 2023.

num_years

The number of years to generate the leasing calendar for. Defaults to 3.

Value

A tibble containing the leasing calendar schedule.