Database Configuration

Functions to retrieve and validate the database configuration.

get_db_config(): retrieves the database configuration from the specified configuration file and returns the configuration as a list. You can optionally specify a key to retrieve a specific configuration value.

validate_db_config(): validates the database configuration to ensure all required configuration keys are present.

Usage

get_db_config(
  key = NULL,
  file = Sys.getenv("R_CONFIG_FILE", pkg_sys("config/config.yml")),
  config = Sys.getenv("R_CONFIG_ACTIVE", "default")
)

validate_db_config(
  cfg,
  arg = rlang::caller_arg(cfg),
  call = rlang::caller_env()
)

Arguments

key

Character string representing the configuration key to retrieve. Defaults to NULL which returns the full configuration list.

file

A character string representing the path to the configuration file.

config

A character string representing the configuration to use from the configuration file.

cfg

List containing the configuration values or a string to the path of a configuration file. If NULL, the default configuration file path will be used.

Value

get_db_config(): A list, or vector (if key is specified), corresponding to the contents of the db configuration key's values (i.e. dbname, host, port, username, and password). If key is specified, the function will return the value corresponding to the specified key.

validate_db_config(): An invisible NULL if the configuration is valid. The function will throw an error if the configuration is invalid.

See also