Connect to GMH Database

Functions to connect, disconnect, and manage the database connection to the GMH PostgreSQL database.

db_connect(): Connect to the GMH PostgreSQL database.

db_disconnect(): Disconnect from the GMH PostgreSQL database.

db_checkout(): Check out a connection from the database connection pool.

db_return(): Return a connection to the database connection pool.

Usage

db_connect(db_config = get_db_config(), user_id = NULL)

db_disconnect(conn)

db_checkout(conn)

db_return(conn)

Arguments

db_config

Database configuration values as a list. The configuration values should include the following keys: host, port, dbname, user, and password at a minimum. If left NULL, will retrieve the configuration via the get_db_config() function.

user_id

An integer representing the user id to use for the database connection. If not provided, the function will attempt to retrieve the user id from the shiny session object. If the function is not running inside a shiny app, the default user id of 0 will be used.

conn

A database connection object.

Value

A pool object representing the connection pool to the GMH PostgreSQL database.

Details

The db_connect() function establishes a connection pool to the GMH PostgreSQL database hosted in Google Cloud SQL using the pool package.

Additionally, the function performs the following:

Sets up automatic disconnect and cleanup of the connection pool when the R or shiny session ends.

Sets the app.user_id database session variable

Stores the connection pool in the shiny session object

Stores the user id in the shiny session object

Sets the db.pool option to the connection pool