Weekly Period

These functions handle calculations related to weekly periods, which are typically used for reporting purposes. A weekly period is defined as a period of seven days.

The following functions are defined:

get_weekly_period(): Returns the start and end dates of the weekly period.

get_weekly_period_start_date(): Returns the start date of the weekly period.

There is no need for a function to calculate the end date of the weekly period, as it is simply the "as of" date.

Usage

get_weekly_period(as_of_date = Sys.Date())

get_weekly_period_start_date(as_of_date = Sys.Date())

Arguments

as_of_date

Date to use as the reference point for deriving other dates from. Defaults to the current system date.

Value

get_weekly_period(): A list containing the start and end dates of the weekly period.

get_weekly_period_start_date(): The start date of the weekly period.

See also

Examples

get_weekly_period()
get_weekly_period_start_date()